Anding a state machine approach to parsing through some of the basic table of contents notes.
Seperating the "patterns" out to be processed in a couple
of different ways, would be useful to start doing some statistical analysis,
but likely need a larger part of the population for that.

First format can be done using regular expressions ( title / author -- title /author)

Second format (and possibly others) of
last name, x. y., person_title. work title1. work title2.--

Are more complicated, trying a state machien approach.
For this pattern it's ParseTwoFSM.pm (finite-state machine)

I'm being a little loose w/ the implementation, can make it
more standardized later if we end up needing more of these.

Currently have it so it can parse in the last name and initials.

Using t/play_with_parser.pl as scaffolding to build up the state machine,
with some dummy functions that work as the unimplemented states.