Several different tweaks
* First, added something to try to "guess" when a title contains
a form of address and hence shouldn't break on a period

Cleaned up some of the formatting of the fields, eliminating spaces
at start and multiple spaces in a row. Probably could clean up the
state machine a bit, but this also helps deal with human error.

Had to do some tweaking of the test cases to make the record
an "ideal" example, since I realize at this point there's
going to be bad data, but I want to get a reasonable
initial parsing that can be fixed later.